merge latest progress

+++ src/core/tools/simulator/task_engine.sim.cpp
@@ -121,7 +121,7 @@ bool sim_semaphore_provider::wait(int timeout_milliseconds)
     else
     {
         // wait success
-        if (timeout_milliseconds == TIME_MS_MAX || dsn_probability() <= 0.5)
+        if (static_cast<unsigned int>(timeout_milliseconds) == TIME_MS_MAX || dsn_probability() <= 0.5)
         {
             _wait_threads.push_back(scheduler::task_worker_ext::get(task::get_current_worker()));
             scheduler::instance().wait_schedule(true, false);
